Fix Decimal type mismatch for latitude and longitude in InMemoryGymsRepository

+++ src/repositories/in-memory/in-memory-gyms-repository.ts
@@ -1,4 +1,5 @@
 import { randomUUID } from 'node:crypto'
+import type { Decimal } from '@prisma/client/runtime/library'
 import type { Gym } from '../../../generated/prisma/client'
 import type { GymCreateInput } from '../../../generated/prisma/models'
 import type { GymsRepository } from '../gyms-repository'
@@ -20,8 +21,8 @@ export class InMemoryGymsRepository implements GymsRepository {
       title: data.title,
       description: data.description ?? null,
       phone: data.phone ?? null,
-      latitude: data.latitude,
-      longitude: data.longitude,
+      latitude: data.latitude as Decimal,
+      longitude: data.longitude as Decimal,
     }
     this.items.push(gym)
     return gym
